requirements.txt
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi
chardet==3.0.4
flow==0.0.1
google==3.0.0
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
numpy==1.19.1
oath==1.4.3
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.1.1
pip==19.0.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
setuptools==40.8.0
six==1.15.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib2-file==0.2.1
urllib2-prior-auth==0.2.0
urllib3==1.25.10
